requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
requests==2.28.1
redis==5.0.1



